Restructured device handling into directory
