+++ src-tauri/src/mesh/device/helpers.rs
@@ -1,0 +1,10 @@
+use std::time::UNIX_EPOCH;
+
+pub fn get_current_time_u32() -> u32 {
+    std::time::SystemTime::now()
+        .duration_since(UNIX_EPOCH)
+        .expect("Could not get time since unix epoch")
+        .as_secs()
+        .try_into()
+        .expect("Could not convert u128 to u32")
+}
