+++ src-tauri/src/mesh/mod.rs
@@ -1,2 +1,2 @@
-pub(crate) mod device;
-pub(crate) mod serial_connection;
+pub mod device;
+pub mod serial_connection;
